python demo service split in separate server and client implementation
