+++ samples/python/minimal_service/minimal_service.pyproj
@@ -21,7 +21,8 @@
     <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="minimal_service.py" />
+    <Compile Include="minimal_service_server.py" />
+    <Compile Include="minimal_service_client.py" />
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
