+++ samples/python/minimal_service/minimal_service_server.py
@@ -27,13 +27,13 @@ def main():
   print("eCAL {} ({})\n".format(ecal_core.getversion(), ecal_core.getdate()))
   
   # initialize eCAL API
-  ecal_core.initialize(sys.argv, "py_minimal_service")
+  ecal_core.initialize(sys.argv, "py_minimal_service_server")
   
   # set process state
   ecal_core.set_process_state(1, 1, "I feel good")
 
-  # create server "Ping"
-  server = ecal_service.Server("DemoServer")
+  # create a server for the "DemoService" service
+  server = ecal_service.Server("DemoService")
 
   # define the server method "foo" function
   def foo_req_callback(method_name, req_type, resp_type, request):
@@ -49,44 +49,11 @@ def main():
   server.add_method_callback("foo",  "string",    "string",    foo_req_callback)
   server.add_method_callback("ping", "ping_type", "pong_type", ping_req_callback)
 
-  # create a client for the "DemoServer" service
-  client = ecal_service.Client("DemoServer")
-
-  # define the client response callback to catch server responses
-  def client_resp_callback(service_info, response):
-    if (service_info["call_state"] == "call_state_executed"):
-      print("'DemoService' method '{}' responded : '{}'".format(service_info["method_name"], response))
-      print()
-    else:
-      print("server {} response failed, error : '{}'".format(service_info["host_name"], service_info["error_msg"]))
-      print()
-
-  # and add it to the client
-  client.add_response_callback(client_resp_callback)
-
-  # match all
-  time.sleep(2.0)
-  
-  # call foo method
-  i = 0
-  while ecal_core.ok() and i < 20:
-    i = i + 1
-    request = bytes("hello foo {}".format(i), "ascii")
-    print("'DemoService' method 'foo' requested with : {}".format(request))
-    client.call_method("foo", request)
+  # idle
+  while(ecal_core.ok()):
     time.sleep(1.0)
 
-  # call ping method
-  i = 0
-  while ecal_core.ok() and i < 20:
-    i = i + 1
-    request = bytes("ping number {}".format(i), "ascii")
-    print("'DemoService' method 'ping' requested with : {}".format(request))
-    client.call_method("ping", request)
-    time.sleep(1.0)
-
-  # destroy client and server
-  client.destroy()
+  # destroy server
   server.destroy()
   
   # finalize eCAL API
